+++ lib/declarative/defaults.rb
@@ -1,4 +1,6 @@
 module Declarative
+  # {Defaults} is a mutable DSL object that collects default directives via #merge!.
+  # Internally, it uses {Variables} to implement the merging of defaults.
   class Defaults
     def initialize
       @static_options  = {}
@@ -19,16 +21,25 @@ module Declarative
       # TODO: allow to receive rest of options/block in dynamic block. or, rather, test it as it was already implemented.
       evaluated_options = @dynamic_options.(name, given_options)
 
-      options = Declarative::Variables.merge( @static_options, handle_array_and_deprecate(evaluated_options) )
-      options = Declarative::Variables.merge( options, handle_array_and_deprecate(given_options) ) # FIXME: given_options is not tested!
+      options = Variables.merge( @static_options, handle_array_and_deprecate(evaluated_options) )
+      options = Variables.merge( options, handle_array_and_deprecate(given_options) ) # FIXME: given_options is not tested!
     end
 
     def handle_array_and_deprecate(variables)
-      wrapped = Hash[ variables.find_all { |k,v| v.is_a?(Array) }.collect { |k,v| [k, Variables::Append(v)] } ]
+      wrapped = Defaults.wrap_arrays(variables)
 
-      warn "[Declarative] Defaults#merge! and #call still accept arrays and automatically prepend those. This is now deprecated, you should replace `ary` with `Declarative::Variables::Append(ary)`."
+      warn "[Declarative] Defaults#merge! and #call still accept arrays and automatically prepend those. This is now deprecated, you should replace `ary` with `Declarative::Variables::Append(ary)`." if wrapped.any?
 
       variables.merge(wrapped)
+    end
+
+    # Wrap arrays in `variables` with Variables::Append so they get appended to existing
+    # same-named arrays.
+    def self.wrap_arrays(variables)
+      Hash[ variables.
+        find_all { |k,v| v.is_a?(Array) }.
+        collect  { |k,v| [k, Variables::Append(v)] }
+      ]
     end
   end
 end